Replace Wintersmith with Gulp
Why? Because Wintersmith was more suited to blog, which this isn't, and
hid too much stuff behind another layer of indirection. It wasn't clear
how to make it do the rather simple things I wanted to do.

All I want is a some static pages with a sensible asset pipeline. It's
more work to get that setup (and maintain?) with Gulp but at least
I understand, and have control over, every step of the process.